pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "defender-cloud-apps-api-client"
version = "0.1.0"
description = "Python client for Microsoft Defender for Cloud Apps API"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
keywords = ["microsoft", "defender", "cloud apps", "security", "api", "mdca", "casb"]
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "Intended Audience :: Information Technology",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "requests>=2.31.0",
    "urllib3>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "types-requests>=2.31.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/defender-for-cloud-apps-api-client"
Documentation = "https://learn.microsoft.com/en-us/defender-cloud-apps/api-introduction"
Repository = "https://github.com/yourusername/defender-for-cloud-apps-api-client"
"Bug Tracker" = "https://github.com/yourusername/defender-for-cloud-apps-api-client/issues"

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
